Add enums folder
