+++ website/common/src/enums/add-events.enum.ts
@@ -1,0 +1,11 @@
+enum Keys {
+  MessageEnqueued = 'MessageEnqueued',
+  UserMessageSent = 'UserMessageSent',
+  UserMessageRead = 'UserMessageRead',
+  MessagesDispatched = 'MessagesDispatched',
+  ProgramChanged = 'ProgramChanged',
+  CodeChanged = 'CodeChanged',
+  DatabaseWiped = 'DatabaseWiped',
+}
+
+export { Keys };
