+++ website/common/src/enums/code.enum.ts
@@ -1,0 +1,6 @@
+enum CODE_STATUS {
+  ACTIVE = 'Active',
+  INACTIVE = 'Inactive',
+}
+
+export { CODE_STATUS };
