+++ website/common/src/enums/program.enum.ts
@@ -1,0 +1,7 @@
+enum InitStatus {
+  SUCCESS = 'success',
+  FAILED = 'failed',
+  PROGRESS = 'in progress',
+}
+
+export { InitStatus };
