+++ website/common/src/enums/rpc-error-code.enum.ts
@@ -1,0 +1,8 @@
+enum RpcErrorCode {
+  GearError = -32602,
+  MethodNotFoundError = -32601,
+  InternalServerError = -32500,
+  UnathorizedError = -32401,
+}
+
+export { RpcErrorCode };
