--- website/common/src/gear/index.ts
@@ -1,1 +1,0 @@
-export * from './gear-events';
