+++ website/common/src/index.ts
@@ -1,5 +1,4 @@
 export * from './logger';
 export * from './interfaces';
+export * from './enums';
 export * from './jsonrpc-errors';
-export * from './kafka-topics';
-export * from './gear';
