+++ website/common/src/interfaces/message.ts
@@ -1,3 +1,5 @@
+import { MESSAGE_READ_STATUS } from '../enums';
+
 interface IMessage {
   id: string;
   destination: string;
@@ -26,22 +28,4 @@ interface IUserMessageReadData {
   reason: MESSAGE_READ_STATUS | null;
 }
 
-enum MESSAGE_TYPE {
-  USER_MESS_SENT = 'UserMessageSent',
-  ENQUEUED = 'Enqueued',
-}
-
-enum MESSAGE_READ_STATUS {
-  OUT_OF_RENT = 'OutOfRent',
-  CLAIMED = 'Claimed',
-  REPLIED = 'Replied',
-}
-
-export {
-  IMessage,
-  IMessageEnqueuedData,
-  IMessagesDispatchedData,
-  IUserMessageReadData,
-  MESSAGE_TYPE,
-  MESSAGE_READ_STATUS,
-};
+export { IMessage, IMessageEnqueuedData, IMessagesDispatchedData, IUserMessageReadData };
